Lab, Final Monster, Quality of Life
- Elements now destroy on collision
- Fixed collision issues with player/ground
- Added two new scenes ( lab, startscene)
- Spaced out buildings
- finalized monster1
- imported finalmonster model
- begun working on labscene

+++ 3d  project/Assets/Scripts/PlayerMovement.cs
@@ -147,7 +147,7 @@ public class PlayerMovement : MonoBehaviour
     {
         health -= damage;
         Debug.Log(health);
-        healthBar.fillAmount = health / 100;
+        healthBar.fillAmount = (float) health / 100f;
     }
     #endregion
 }
